rack_query: move to ThreeScale::Client namespace
